# specs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/teams": {
      "get": {
        "operationId": "TeamsController_findAll",
        "parameters": [
          {
            "name": "clubId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of teams",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TeamListItemDto"
                  }
                }
              }
            }
          }
        },
        "summary": "Get all teams",
        "tags": [
          "teams"
        ]
      },
      "post": {
        "operationId": "TeamsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTeamDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Team created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TeamResponseDto"
                }
              }
            }
          }
        },
        "summary": "Create a new team",
        "tags": [
          "teams"
        ]
      }
    },
    "/teams/{id}": {
      "get": {
        "operationId": "TeamsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Team details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TeamResponseDto"
                }
              }
            }
          }
        },
        "summary": "Get a team by ID",
        "tags": [
          "teams"
        ]
      },
      "patch": {
        "operationId": "TeamsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTeamDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Team updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TeamResponseDto"
                }
              }
            }
          }
        },
        "summary": "Update a team",
        "tags": [
          "teams"
        ]
      },
      "delete": {
        "operationId": "TeamsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Team deleted"
          }
        },
        "summary": "Delete a team",
        "tags": [
          "teams"
        ]
      }
    },
    "/teams/club/{clubId}": {
      "get": {
        "operationId": "TeamsController_findTeamsByClubId",
        "parameters": [
          {
            "name": "clubId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of teams for the club",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TeamListItemDto"
                  }
                }
              }
            }
          }
        },
        "summary": "Get teams by club ID",
        "tags": [
          "teams"
        ]
      }
    },
    "/teams/{teamId}/assign-player/{playerId}": {
      "post": {
        "operationId": "TeamsController_assignPlayerToTeam",
        "parameters": [
          {
            "name": "teamId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "playerId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Player assigned to team",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TeamResponseDto"
                }
              }
            }
          }
        },
        "summary": "Assign a player to a team",
        "tags": [
          "teams"
        ]
      }
    },
    "/teams/{teamId}/remove-player/{playerId}": {
      "delete": {
        "operationId": "TeamsController_removePlayerFromTeam",
        "parameters": [
          {
            "name": "teamId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "playerId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Player removed from team",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TeamResponseDto"
                }
              }
            }
          }
        },
        "summary": "Remove a player from a team",
        "tags": [
          "teams"
        ]
      }
    }
  },
  "info": {
    "title": "PlayerTracker API",
    "description": "API pour la plateforme de suivi physique et mental des athlètes",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "ClubNestedDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "Club ID",
            "example": 1
          },
          "name": {
            "type": "string",
            "description": "Club name",
            "example": "FC Barcelona"
          },
          "address": {
            "type": "object",
            "description": "Club address",
            "example": "123 Main St",
            "nullable": true
          },
          "phone": {
            "type": "object",
            "description": "Phone number",
            "example": "+34 123 456 789",
            "nullable": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "Creation timestamp",
            "example": "2024-01-01T00:00:00Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "Last update timestamp",
            "example": "2024-01-15T12:00:00Z"
          }
        },
        "required": [
          "id",
          "name",
          "address",
          "phone",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "CountDto": {
        "type": "object",
        "properties": {
          "players": {
            "type": "number",
            "description": "Number of players",
            "example": 15
          }
        },
        "required": [
          "players"
        ]
      },
...
      "TeamListItemDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "Team ID",
            "example": 1
          },
          "name": {
            "type": "string",
            "description": "Team name",
            "example": "First Team"
          },
          "description": {
            "type": "string",
            "description": "Team description",
            "example": "Senior football team",
            "nullable": true
          },
          "logoUrl": {
            "type": "string",
            "description": "Team logo URL",
            "example": "https://example.com/logo.png",
            "nullable": true
          },
          "sportId": {
            "type": "number",
            "description": "Sport ID",
            "example": 1
          },
          "clubId": {
            "type": "number",
            "description": "Club ID",
            "example": 1
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "Creation timestamp",
            "example": "2024-01-01T00:00:00Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "Last update timestamp",
            "example": "2024-01-15T12:00:00Z"
          },
          "club": {
            "description": "Club details",
            "allOf": [
              {
                "$ref": "#/components/schemas/ClubNestedDto"
              }
            ]
          },
          "_count": {
            "description": "Counts",
            "allOf": [
              {
                "$ref": "#/components/schemas/CountDto"
              }
            ]
          }
        },
        "required": [
          "id",
          "name",
          "sportId",
          "clubId",
          "createdAt",
          "updatedAt",
          "club",
          "_count"
        ]
      },
      "PlayerNestedDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "Player ID",
            "example": 1
          },
          "firstName": {
            "type": "string",
            "description": "First name",
            "example": "John"
          },
          "lastName": {
            "type": "string",
            "description": "Last name",
            "example": "Doe"
          },
          "nickName": {
            "type": "string",
            "description": "Nickname",
            "example": "Johnny",
            "nullable": true
          },
          "isActive": {
            "type": "boolean",
            "description": "Is active",
            "example": true
          },
          "positionId": {
            "type": "number",
            "description": "Position ID",
            "example": 1,
            "nullable": true
          }
        },
        "required": [
          "id",
          "firstName",
          "lastName",
          "nickName",
          "isActive",
          "positionId"
        ]
      },
      "QuestionnaireNestedDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "Questionnaire ID",
            "example": 1
          },
          "title": {
            "type": "string",
            "description": "Title",
            "example": "Weekly Check-in"
          },
          "status": {
            "type": "string",
            "description": "Status",
            "example": "DRAFT",
            "enum": [
              "DRAFT",
              "ACTIVE",
              "ARCHIVED"
            ]
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "Creation timestamp",
            "example": "2024-01-01T00:00:00Z"
          }
        },
        "required": [
          "id",
          "title",
          "status",
          "createdAt"
        ]
      },
      "TeamResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "Team ID",
            "example": 1
          },
          "name": {
            "type": "string",
            "description": "Team name",
            "example": "First Team"
          },
          "description": {
            "type": "string",
            "description": "Team description",
            "example": "Senior football team",
            "nullable": true
          },
          "logoUrl": {
            "type": "string",
            "description": "Team logo URL",
            "example": "https://example.com/logo.png",
            "nullable": true
          },
          "sportId": {
            "type": "number",
            "description": "Sport ID",
            "example": 1
          },
          "clubId": {
            "type": "number",
            "description": "Club ID",
            "example": 1
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "Creation timestamp",
            "example": "2024-01-01T00:00:00Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "Last update timestamp",
            "example": "2024-01-15T12:00:00Z"
          },
          "club": {
            "description": "Club details",
            "allOf": [
              {
                "$ref": "#/components/schemas/ClubNestedDto"
              }
            ]
          },
          "players": {
            "description": "Players",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PlayerNestedDto"
            }
          },
          "questionnaires": {
            "description": "Questionnaires",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/QuestionnaireNestedDto"
            }
          }
        },
        "required": [
          "id",
          "name",
          "sportId",
          "clubId",
          "createdAt",
          "updatedAt",
          "club",
          "players",
          "questionnaires"
        ]
      },
      "CreateTeamDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Nom de l'équipe",
            "example": "Équipe première",
            "minLength": 1,
            "maxLength": 100
          },
          "description": {
            "type": "string",
            "description": "Description de l'équipe",
            "example": "Équipe senior masculine de football",
            "maxLength": 500
          },
          "logoUrl": {
            "type": "string",
            "description": "URL du logo",
            "example": "https://example.com/team-logo.png"
          },
          "clubId": {
            "type": "number",
            "description": "ID du club",
            "example": 1,
            "minimum": 1
          },
          "sportId": {
            "type": "number",
            "description": "ID du sport",
            "example": 1,
            "minimum": 1
          }
        },
        "required": [
          "name",
          "clubId",
          "sportId"
        ]
      },
      "UpdateTeamDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Nom de l'équipe",
            "example": "Équipe première",
            "minLength": 1,
            "maxLength": 100
          },
          "description": {
            "type": "string",
            "description": "Description de l'équipe",
            "example": "Équipe senior masculine de football",
            "maxLength": 500
          },
          "logoUrl": {
            "type": "string",
            "description": "URL du logo",
            "example": "https://example.com/team-logo.png"
          },
          "clubId": {
            "type": "number",
            "description": "ID du club",
            "example": 1,
            "minimum": 1
          },
          "sportId": {
            "type": "number",
            "description": "ID du sport",
            "example": 1,
            "minimum": 1
          }
        }
      }
    }
  }
}